Rename isCygwin() to isXterm()
Strictly speaking this method does not check for Cygwin (as it e.g. does
not check to be on Windows) but only for XTerm (which in conjunction
with being on Windows is a sign of running under Cygwin).

Simplify the code a bit to get rid of the explicit null check long the
way.